리다이렉션 손보는 중

+++ exec_child.c
@@ -6,40 +6,50 @@
 /*   By: jeseo <jeseo@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/13 15:40:48 by jeseo             #+#    #+#             */
-/*   Updated: 2023/03/13 21:15:37 by jeseo            ###   ########.fr       */
+/*   Updated: 2023/03/14 18:24:48 by jeseo            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-int	stdio_to_pipe(t_info *info, int i, int fd[], int temp_fd)
+int	stdio_to_pipe(t_cmd *cmd_line, t_pipe_index index, int pipes)
 {
-	if (info->pipes == 0)
+	FILE *out;
+	char *temp;
+	static int i;
+
+	i++;
+	temp = ft_itoa(i);
+	out = fopen(temp, "w");
+	fprintf(out, "fd[0]: %d fd[1]: %d\n", index.fd[0], index.fd[1]);
+	fclose(out);
+	free(temp);
+	if (pipes == 0)
 	{
 		return (0);
 	}
-	if (i == 0)
+	if (index.i == 0)
 	{
 		//첫 번째 커맨드는 입력은 STDIN(0)으로 내버려 둔다.
-		close(fd[0]);
-		dup2(fd[1], STDOUT_FILENO);
-		close(fd[1]);
+		close(index.fd[0]);
+		dup2(index.fd[1], STDOUT_FILENO);
+		close(index.fd[1]);
 	}
-	else if (i == info->pipes)
+	else if (index.i == pipes)
 	{
 		//STDIN(0)을 temp(이전 fd[0])으로
 		//마지막 커맨드는 STDOUT(1)을 연결하지 않고 종료.
-		dup2(temp_fd, STDIN_FILENO);
-		close(temp_fd);
+		dup2(index.prev_pipe_read, STDIN_FILENO);
+		close(index.prev_pipe_read);
 	}
 	else
 	{
 		//0을 temp(이전 fd[0])으로
 		//1을 pipe의 fd[1]로
-		dup2(temp_fd, STDIN_FILENO);
-		dup2(fd[1], STDOUT_FILENO);
-		close(temp_fd);
-		close(fd[1]);
+		dup2(index.prev_pipe_read, STDIN_FILENO);
+		dup2(index.fd[1], STDOUT_FILENO);
+		close(index.prev_pipe_read);
+		close(index.fd[1]);
 	}
 	return (0);
 }
@@ -54,12 +64,14 @@ int	handle_redirection(t_arg_deque *redirections)
 	red = redirections->head;
 	while (red != NULL)
 	{
+		fd = -1;
 		if (red->special == HEREDOC || red->special == REDIRECT_IN)
 		{
 			fd = open(red->arg, O_RDONLY); // 어차피 acess 함수에서 리드 권한으로 열 건지 쓰기 권한으로 열 건지 체크를 해야할 것이다.
 			if (fd < 0)
 				break ;
-			unlink(red->arg);
+			if (red->special == HEREDOC)
+				unlink(red->arg);
 			dup2(fd, STDIN_FILENO);
 			close(fd);
 		}
@@ -89,15 +101,19 @@ int	handle_redirection(t_arg_deque *redirections)
 }
 
 
-int child_process_run(t_info *info, int i, int fd[], int temp_fd)
+int	child_process_run(t_cmd *cmd_node, t_pipe_index index, t_info *info)
 {
-	printf("child process 입니다.\n");
-	stdio_to_pipe(info, i, fd, temp_fd);
-	handle_redirection(info->cmds->head->redirections); // command 노드가 NULL이면 어디서 처리가 되는가?
+	stdio_to_pipe(cmd_node, index, info->pipes);
+	handle_redirection(cmd_node->redirections); // command 노드가 NULL이면 어디서 처리가 되는가?
+	printf("exec: %s\n", cmd_node->commands_args[0]);
+	if (exec_builtin(cmd_node->commands_args) == 0)
+	{
+		exit(EXIT_SUCCESS);
+	}
+	printf("no빌트\n");
 	//command 찾기.
-	execve(info->cmds->head->commands_args[0], info->cmds->head->commands_args, info->envp_bash);
+	execve(cmd_node->commands_args[0], cmd_node->commands_args, info->envp_bash);
 	write(2, "exec 실행 오류.. 어떻게 해야하는데..\n", 49);
-	exit(1);
 	//전역변수 에러 코드 변환해주기.
 	return (0);
 }
